+++ spec/requests/as_account/account_spec.rb
@@ -97,6 +97,21 @@ describe Yt::Account, :device_app do
     end
   end
 
+  describe '.video_groups' do
+    let(:video_group) { $account.video_groups.first }
+
+    specify 'returns the first video-group created by the account' do
+      expect(video_group).to be_a Yt::VideoGroup
+      expect(video_group.title).to be_a String
+      expect(video_group.item_count).to be_an Integer
+      expect(video_group.published_at).to be_a Time
+    end
+
+    specify 'allows to run reports against each video-group' do
+      expect(video_group.views).to be
+    end
+  end
+
   describe '.upload_video' do
     let(:video_params) { {title: 'Test Yt upload', privacy_status: 'private', category_id: 17} }
     let(:video) { $account.upload_video path_or_url, video_params }